Retain choice parameters in ChoiceTransform (#565)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/565

- Changes the behavior of `ChoiceTransform` so that it does not automatically transform choice parameters to integer range parameters.
- Identifies choice parameters properly as (ordinal or categorical) choice parameters inside  `extract_search_space_digest` and pass this information to the model via `SearchSpaceDigest`.
- `TaskEncode` now converts all task parameters to a `ChoiceParameter` of int type with `is_task=True`.

`ChoiceTransform` now retains `ChoiceParameter`s (ordinal and categorical) as `ChoiceParameter`s. For numeric `ChoiceParameter`s (i.e. discrete ordinal parameters), the values are normalized and embedded into `[0, 1]` so as to maintain the relative spacing. For categorical parameters this uses integer encoding.

What this allows us to do is use the standard modeling approach for ordinal (discretized numerical) parameters (without converting them to RangeParameters in the botorch model), and the categorical kernel for categorical parameters.

To this end D27855811 adds another (optional) `discrete_choices` field to `SearchSpaceDigest` that carries the possible discrete choices so we can use them during the optimization without having to round anything

There is a slight issue with these lookup tables since if you stick a python float into a torch tensor, do stuff with it, and then return that tensor as a list of python floats there is a slight discrepancy in the numerical value. To this end I implemented a `ClosestLookupDict` for the inverse transforms, which allows to deal with this issue quite elegantly, and also pretty efficiently (by always maintaining a sorted set of keys and doing nearest-point lookup by using fast sorted search).

The functionality of `OrderedChoiceEncode` is left as is (still converts to integer `RangeParameter`), so this is fully backwards-compatible.

Reviewed By: sdaulton

Differential Revision: D27150313

fbshipit-source-id: 9598e618c878df6d1f6a92660526643ecaa24b9a