Convenience function for tile plot of observations
Summary:
Right now if you want to make a tile plot of all of the observations on an experiment, you have to first fit a TS model:
```
m_ts = Models.THOMPSON(
    data=experiment.fetch_data(),
    search_space=experiment.search_space,
    experiment=experiment,
)
render(tile_fitted(m_ts))
```
I find myself doing this fairly often so I decided to make a convenience function for doing it "tile_observations" that you can just pass in the experiment and it does the rest.

I also added a functionality to the Pareto plot that allows you to filter to a specified set of arms, because that is part of the ParallAx flow.

Reviewed By: stevemandala

Differential Revision: D28207299

fbshipit-source-id: 8ca25007c46a09d9b21dc4baf31f91aaa5c80cf6