make reference point optional (#601)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/601

This diff does two things: 1) if a user creates a `MultiObjectiveTorchModelbridge` and calls `gen` without specifying objective thresholds, then we infer the objective thresholds within the `MultiObjectiveBotorchModel`, generate candidates, and return the inferred objective thresholds in `gen_metadata`. 2) it adds a `infer_objective_thresholds` method to the `MultiObjectiveTorchModelbridge`, which can be used to infer objective thresholds without generating candidates.

This refactors the Base `Modelbridge.gen` and `ArrayModelbridge._gen` methods and to apply transformations within a utility function.

Note that this method returns ObservationData. If the user wants to plot outcomes with objective thresholds, the user would have to create the ObjectiveThresholds and set the objective thresholds on the optimization config.

Reviewed By: Balandat

Differential Revision: D28163744

fbshipit-source-id: 9a674cc51631ab85de23a325202445c3a06620aa