untangle trials from different sources
Summary:
Failure rate calculations are currently affected by the source of a trial, which can be warm-started, from a previous fblearner retry, or user-supplied. specifically, we want to count trials run by the current scheduler, as well as trials from previous fblearner retries of the same sweep. However, "user-supplied" and "warm-start" trials should _not_ affect the failure rate calculation.

This diff introduces a trial._properties["source"] which is intended to track which trials belong to which group above, and to have the appropriate effect on the failure rate calculation. This is done by explicitly reducing `scheduler._num_preexisting_trials` by the number of trials from previous fblearner retries in `_get_scheduler_and_config()`.

Reviewed By: lena-kashtelyan

Differential Revision: D29180056

fbshipit-source-id: d4d75035dfd24449233864ab5b42e8008b6bf7f2