Fix for missing arms
Summary:
See full context here:  T93796084

The tldr; here is that our experiment saving relies on a tricky function called "copy db ids". The way this function is used is this:

1. We have a user facing object Experiment that we want to save
2. We convert this to a SQLAlchemy object, call it SQAExperiment
3. We merge SQAExperiment into the session (i.e. add/update the DB as needed)
4. The result of step 3 is that we get back a new SQLAlchemy object, call it SQAExperimentNew, that has database IDs attached to all the new objects it created
5. We now want to annotate the original user facing Experiment object with these db ids also! So we use `copy_db_ids` to transfer these db ids over

`copy_db_ids` is a difficult operation to do correctly. At its core, it should be a relatively simply recursive operation that just copies a `db_id` field from one object to another. But because these objects have lots of nested fields and relationships, there are exceptions ...

I'm happy to explain further, but for the purposes of unblocking things, sufficient to say that this diff handles an important edge case where there are two copies of an identical arm on an experiment (one is attached to one trial and one attached to another).

I have a task open already to revisit the logic here and see if it can be simpler (T88027346).

Differential Revision: D29280391

fbshipit-source-id: 1eb390e8f2a34a1798960f4faac69a94d46a41e1