update default partitioning selection heuristic (#846)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/846

We now have a very fast non-dominated partitioning that is much more scalable with the number of objectives.

Using 100 data points on the PF (a very large PF):
For 3 objectives, FastNondominatedPartitioning (exact) is faster. For 4 objectives, FastNondominatedPartitioning (exact) is competitive with the approximate heuristic. For 5 objectives, FastNondominatedPartitioning is a good deal slower (still not that bad).

Reviewed By: Balandat

Differential Revision: D29353926

fbshipit-source-id: a12345a6a6f4a3427af8e4e8bc46a78fc54b5c02