Fix MapData storage error
Summary:
MapData is not currently storing properly, see NB here: N932571. This is partly because we added MapData to the [core Ax SQAConfig](https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/ax/ax/storage/sqa_store/sqa_config.py?commit=988ff07f92fa&lines=65) but [not the FB-specific one](https://fburl.com/diffusion/vw0o7jdm).

But really, we shouldn't need to add it to either config, since  all of our user-facing Data classes map to SQAData, so we can make a simple change in Encoder and not worry about the SQAConfig.

In the future, if we start storing data in different tables or formats, we'll need to change this, but that's probably a ways off.

Reviewed By: Balandat

Differential Revision: D29516718

fbshipit-source-id: d08c9d30e88a4189ca7643d8c3c90166b1536103