Add reason input to Runner.stop and return stop metadata
Summary:
Changes the signature of `Runner.stop` to have an additional `reason` argument for specifying the reason for stopping. This can be used by runners to inform dispatch systems of why a trial run has been stopped.

Also makes `stop` return a dictionary of metadata that can be used downstream (as the `run` method already has). This can be used to store information obtained from the backend system in the process of stopping the trial.

Reviewed By: ldworkin

Differential Revision: D29351464

fbshipit-source-id: 09418fcc729c3c736fc6fdd1e12849ea3425d3f4