Upstream `suppress_storage_errors` into `WithDBSettingsBase` for use in `Scheduler`
Summary:
For the sweep, we want to run the sweep even if we run into issues with storage; in general, a setting like this will be useful for many `Scheduler` applications.

We'll probably want a special alert on these errors, since they really shouldn't happen, but they also should not fail a sweep when the do (e.g. for a transient reason or due to a bug). Currently looking into how we can make that happen to add to the roadmap

Reviewed By: bernardbeckerman

Differential Revision: D28271835

fbshipit-source-id: f8f7a8e184bb8739cb5d3f9d5342548c9613a504