Reverse save order to fix tutorial
Summary:
Serializing and deserializing `gs` dissociates it with `experiment`, which causes the [stack trace in our docs](https://ax.dev/tutorials/generation_strategy.html#3B.-SQL-storage). Reversing the oder of the SQL and JSON block fixes it.  We could alternatively overwrite
```
experiment = gs._experiment
```
at the end of the JSON block if we feel the order is important.

Reviewed By: ldworkin

Differential Revision: D29702679

fbshipit-source-id: 28f890400f39caec2d9d1927ff711418b58a5d46