remove feature importance plot from standard_plots
Summary: The purpose of this plot from a user perspective is to answer the question of which variables have the most influence over the objective. Per discussions with David, the current implementation of feature_importances uses the [length scale hyperparameter](https://fburl.com/diffusion/hjteuk7h) as a proxy for importance, which will likely mislead the aforementioned user.

Reviewed By: ldworkin

Differential Revision: D29805304

fbshipit-source-id: 48bc921d8456ee9a1a7f5c150124a3aa3d34c405