surface trial._properties in exp_to_df output
Summary: See title. As of this diff, `_properties["source"]` is an empty string when the source is unknown _or_ when the source is the current retry of the current sweep. i want to discuss what the expected behavior should be and follow this up per the result of that discussion.

Reviewed By: lena-kashtelyan

Differential Revision: D29830945

fbshipit-source-id: 1fb3c7958969edc035199ff12c99d32800c8e4f6