Add overwrite_existing_data class property to Metric
Summary:
Adds a class property to Metric that indicates whether we should overwrite previously cached data when attaching new data.

This property should default to False, but should be True for metrics that are available while running, and will report a full set of results every time (e.g. 0 to N+1 rather than just N - N+1). The Curve metric is the best example of this.

Reviewed By: lena-kashtelyan

Differential Revision: D29802638

fbshipit-source-id: d8902e0b9a7338b44f074591d574773183aba4d2