Use `combine_with_last_data` instead of `merge_across_timestamps`
Summary:
## High level context
The goal of this diff (and the subsequent one in the stack) is to support the use case of *incremental* partial results. i.e. a metric, or a system, that returns results N to N+1 at each data fetch, rather than 0 to N+1. Currently, we don't have any use cases that fit this description, but the idea is to be forward looking and make sure we're well set up to support both.

## How this is currently handled
Right now, this use case is supported via the Service API, using the method `update_running_trial_with_intermediate_data`. This will attach a new dataframe with the partial results. After this, you can use `lookup_data(merge_across_timestamps=True)` to fetch all of the partial results, merged into a single dataframe.

## What this diff does instead
This diff implements an alternate approach. Instead of attaching a new dataframe with each new partial result, we use `attach_data(combine_with_last_data=True)` to merge the partial result into the existing dataframe. Thus, we don't need to use the `merge_cross_timestamps` argument at lookup time, and can reap this.

Reviewed By: lena-kashtelyan

Differential Revision: D29836373

fbshipit-source-id: 69e410bffe38028e24b930f4a023e5917962333c