Change metric->metric relationship to speed up loading
Summary:
Our MultiObjectives, ScalarizedObjectives, and ScalarizedOutcomeConstraints are all stored in the database as follows:

* One wrapper "parent" metric that just acts as a placeholder
* Multiple "children" metrics, all of which point to this parent

When we load experiments from our DB, given *any* metric object, we currently eagerly query to see if it has any children. We do this even for non-MOO metrics!

In this diff, we change the behavior of this metric parent -> metric children relationship to be *lazy*. This means we won't query any metric for its children. But, when we decode the SQL class to its user facing counterpart, if we come across a wrapper parent metric, we then reconnect to the DB and fetch its children.

So load time for MOO experiments should be mostly unchanged, but load time for non-MOO experiments should be really sped up.

Reviewed By: stevemandala

Differential Revision: D29865149

fbshipit-source-id: 617e524b70c5fee098e77c3dd7ca5626e35be3f0