Make `get_standard_plots` take in model instead of generation strategy; retrain model in ParamSweepScheduler before calling
Summary: The ModelBridge object stored in generation_strategy.model is currently instantiated at the time of arm creation, so that the model interpolation results as well as the model training_data will not incorporate any results from arms that complete after the last arm (or arms if max_pending_trials > 1) is generated. This re-instantiates/retrains the model before creating any plots that use it.

Reviewed By: lena-kashtelyan

Differential Revision: D30072955

fbshipit-source-id: db2cff93e6d668784f8bc03f3933b8e560b29a36