pre-existing candidate arms should not be counted toward total number of runs
Summary:
Currently in the `Scheduler`, candidate arms that are pre-attached by the user but not yet run are not counted toward the total number of runs to execute. This can be confusing to users who may employ, e.g., the `max_trials` arg to `Scheduler.run_n_trials`, to respect budgetary constraints.

This diff changes this behavior such that any trial that still must be run/executed to produce results is counted toward the total number of trials to be run (e.g., toward `max_trials` in `Scheduler.run_n_trials`)

Reviewed By: lena-kashtelyan

Differential Revision: D30259906

fbshipit-source-id: f80356e0a7e19f5ed56f36b3c355ff0f9e32a263