Solutions from ax.optimize don't satisfy parameter_constraints
Summary:
`parameter_constraints` is not propagated correctly to `SearchSpace` whenever we run an experiment through the `ax.optimize` entry point (like in the tutorial in the official documentation).

The culprit is `Optimization.with_evaluation_function`, which takes as a parameter `parameter_constraints`, but it forgets to pass it to the `make_experiment` factory function.

Updated the unittests to catch the offending behaviour.

Reviewed By: stevemandala

Differential Revision: D30278157

fbshipit-source-id: 309485b64947ee7d54428862865593a93212eb0c