Fix two storage issues that were affecting Waveguide
Summary:
We've been encountering occassional storage timeouts in Waveguide due to the fact that, for some reason, storage updates were deleting arms and then re-inserting them. The re-insert was pre-specifying the primary key, which is unideal and slows down the insert, resulting in timeouts.

While I don't know why the re-insert was pre-specifying the primary key, the reason that the arms were being deleted in the first place was because their `db_id` was not properly being set in `copy_db_ids`, due to the fact that the arm objects are nested in an `ArmWeight` object, which did not inherit from `Base`, and was therefore being skipped over [here](https://fburl.com/diffusion/q48fsxs9). Changing `ArmWeight` to inherit from Base solves this problem :)

While investigating this, I also noticed that `ObjectiveThresholds` were also being deleted and re-inserted. While this didn't seem to be causing the timeouts, I decided to fix this as well. The issue here is that `db_id` was being set properly, but it was being overridden when we did a clone [here](https://fburl.com/diffusion/p1dd86qi). By changing the order of when we set the `db_id`, we can fix this as well. It's not the most elegant solution but I can't think of anything better right now.

Reviewed By: lena-kashtelyan

Differential Revision: D30158506

fbshipit-source-id: 7e603dcf18a681c071751893811a7599ca0c2ed5