Validate that objective thresholds match objectives if specified
Summary:
Currently we get an error from modelbrige when some but not all objective thresholds are specified. This addresses that issue by raising a similar error earlier in the stack.

I'm not sure if this is the right move –– perhaps the right move is to infer just the missing objective? Not sure about that so will defer to sdaulton. I think the current setup is wrong in either case, since we do error, but during candidate generation, so if we do want to infer just the missing objective thresholds, we'll need to get rid of the current error in modelbridge utils.

Reviewed By: sdaulton

Differential Revision: D30675593

fbshipit-source-id: 39d0db3c63a63885df0a25b700a2f762cf4348c9