limit Sobol budget to 20% of num_trials in some cases
Summary:
Discussed this with dme65 and it seems like a good change.

As new methods allow users to explore higher-dimensional search spaces, we are testing the [rule-of-thumb in `choose_generation_strategy`](https://fburl.com/diffusion/4cjhgszl) that performs a number of initialization (Sobol) trials equal to:

```
max(5, len(search_space.parameters))
```

For very high-dimensional search spaces, this rule-of-thumb can result in searches devoting most or all of their budget to Sobol search.

This diff proposes an altered rule in the case that `num_trials` is provided:

```
max(5, min(len(search_space.parameters), num_trials // 5))
```

That is, perform a number of initialization trials equal to `n_parameters` or 20% of `num_trials`, whichever is smaller, but always perform at least 5.

Reviewed By: dme65

Differential Revision: D30674682

fbshipit-source-id: f19b272fd4b0008ea643450b4504f08d3f67e6c3