Convert MapData to Data in most places, except in Early Stopping Strategy
Summary:
The general idea of this diff is, whenever we have a MapData object, in most parts of the codebase, we actually only care about the latest values of the map keys. e.g. if we have values for multiple timestamps, GPEI only cares about the latest value, get_metric_mean only cares about the latest value, etc.

Thus, when we call exp.lookup_data, let's default to "flattening" MapData into a standard Data and only keeping the rows with the most recent map key values.

The only exception should be when we're within an EarlyStoppingStrategy, when we really do need all the rows, so we can just override the boolean flag there.

Reviewed By: lena-kashtelyan

Differential Revision: D30701778

fbshipit-source-id: c312ce24e461000186fdd99afab17f668c31ed87