sort ChoiceParameter values when defaulting to is_ordered
Summary: Ordered `ChoiceParameters` are [mapped to consecutive ints](https://fburl.com/code/l57t91by) based on the order in which the choices are provided by the user, regardless of the `ParameterType`. For example, an INT parameter `[2, 1, 9]` will be mapped to `[0, 1, 2]`, respectively. This may be desired behavior, so we don't want to override this ordering in all cases, however if Ax is setting `is_ordered = True` via its default behavior, we should assume the ordinality implied by the choices' values, not their order.

Reviewed By: lena-kashtelyan

Differential Revision: D30728479

fbshipit-source-id: daa29f725e3592a17a8c35c8ca6f1e80a1be2c99