Preparation for using polling and running multiple trials via `Runner`
Summary:
`scheduler.py` is organized in a way that separates methods, for which we encourage user-overriding and for which we don't endorse it (overriding the latter type of methods could break core scheduler logic in hard-to-debug ways). However, that organization slipped a bit with mine and others' work; this diff restores it.

It also makes the scheduler no longer an abstract class, as it's only abstract method: `poll_trial_status`, becomes required in the next diff: D31031586.

Reviewed By: bernardbeckerman

Differential Revision: D31031590

fbshipit-source-id: 8415986aa8b344626cf5beb0fa2d65b575331979