Set sort_values for SearchSpaceToChoice transform
Summary:
The `hypervolume_by_trial` utility in parallax ultimately uses the SearchSpaceToChoice transform to encode the set of observed arms into a choice parameter, so that we can define a model on top of an arbitrary search space.

The change to sorting choice parameter values in D30935826 (https://github.com/facebook/Ax/commit/5f5c5139c74e0eefb8d2983f911e21baeb3a5e43) made it so that you get a warning anytime you define a choice parameter over strings without specifying if it should be sorted or not. So now running the parallax utility produces a long dump of
```
[WARNING 09-24 14:59:32] ax.core.parameter: `sort_values` is not specified for `ChoiceParameter` "arms". Defaulting to `False` for parameters of `ParameterType` STRING. To override this behavior (or avoid this warning), specify `sort_values` during `ChoiceParameter` construction." during hypervolume_by_trial
```
one warning for every trial in the experiment...

This avoids the warning by specifying the sorting.

reviewsmalldiffmmkay

Reviewed By: abbas-kazerouni

Differential Revision: D31210310

fbshipit-source-id: cc52458faa820bc31987ae710ec88803a8688a96