Add validation against constraints on choice and fixed params
Summary:
Previously we would allow choice and fixed parameters to be constrained, which caused confusing errors down the pipeline (If you constrain on a fixed parameter, it is removed in the remove_fixed transform, and you get this error: N997657, if you constrain on a choice parameter, it is renamed and converted to a range parameter in the one_hot transform, and you get this error: N997657). See task for more info.

Now we will raise an exception with a helpful message on experiment creation.

The docstring for create_experiment has also been changed to reflect these restrictions.

Reviewed By: ldworkin

Differential Revision: D31210729

fbshipit-source-id: 26b6e724ea9951375a7de239052f17be1b735773