Prevent reduced-state trials overwriting full-state ones during updates in DB
Summary:
## Problem:
Some generator runs in an experiment are missing model kwargs and model state fields; if such a generator run is the last one in the DB and the generation strategy is in a stateful phase, reloading the generation strategy fails, since the model will need those kwargs and state.

## How we get into that state:
- user runs an experiment with scheduler for a while, it has running trials,
- user stops their kernel/it crashes,
- user resumes their experiment
  - for this, we load the experiment with reduced state (so loaded generator runs attached to experiment don't have model kwargs and state (in DB they exist though);
   - on generation strategy, last loaded generator run is loaded with model kwargs and state, so we are able to reload the generation strategy at this point),
- when the previously running trials complete, we update them to a reduced-state version that doesn't have model kwargs and state (since on the experiment all generator runs are reduced-state, including the one that was last during experiment and GS reloading,
- immediately after that update, user stops their kernel/it crashes,
- now last generator run doesn't have model kwargs and state associated with it in the DB.

Note that this can only happen when trials were created (and marked running) before a resumption, but marked completed (and updated without model kwargs and state) after that resumption. So only this batch of started-before-resumption and completed-after-it trials is affected by this.

 ---

This diff solves the problem by preventing the reduced-state fields from being overwritten with null values

Reviewed By: ldworkin

Differential Revision: D31091947

fbshipit-source-id: c6e1e0fdd5074c430557c13912a84a582800574e