Remove exclude_fields from serialize_init_args
Summary: Getting rid of some extra complexity in the codebase. We initially excluded certain fields from serialization into the metric `properties` because these fields are already stored in their own separate columns in the DB, so it was redundant to store them twice. But really it doesn't matter much to store them twice, and it saves us the headache of dealing with this exclude_fields logic.

Reviewed By: bernardbeckerman

Differential Revision: D31241381

fbshipit-source-id: d0980681775ca6128abb1639dda6822565522c4b