Add option to trial encoder and generation_strategy encoder to save reduced state generator run
Summary:
Added an option to `trial_to_sqa` that when flipped will cause us to elide model state related fields from each generator run.

For context: Waveguide experiments are very large and their models are stateful; we wind up saving much more data than in a typical Ax experiment and much of it is not needed. Specifically, we really only need the model attributes on final run of a batch (in order to resume if the experiment is stopped for some reason) not the attributes for every individual model. See the task for more detail.

Reviewed By: ldworkin

Differential Revision: D31405679

fbshipit-source-id: e35307f1dd598f71c569db23a2b05ff7c8d67769