Add option to store model_kwargs only on last generator run of a batch of trials
Summary: Uses the encoders provided by previous diff to provide an optional flag allowing the user to specify whether or not they would like to save all instances of model state (default) or only those attached to the last generator run on a batch of trials. This is typically all the end user needs and saves a significant amount of space for experiments at Waveguide's scale.

Reviewed By: lena-kashtelyan

Differential Revision: D31408686

fbshipit-source-id: fab6e4857d350cbc6ac8b6c23a9c23745364b05f