Fallback to hit-and-run sampler on rejection sampling errors
Summary:
Modifying sampler to resort to polytope sampler in cases where the rejection sampler fails.

Adding a try-catch block to the gen method of the RandomModel class to resort to the polytope sampler upon encountering a search space exhausted exception.  This is optionally performed based on the added fallback_to_sample_polytope argument.

Adding the  fallback_to_sample_polytope argument to the implementation of the gen method in the RandomModelBridge class.  To avoid requiring this parameter in the ModelBridge base class and avoid repeated code for the gen method in the RandomModelBridge class, the code changes the gen method in ModelBridge to rely on helper functions which are then leveraged by the RandomModelBridge implementation.

Reviewed By: lena-kashtelyan

Differential Revision: D31431838

fbshipit-source-id: e16123cf5f115a77dfbad7d2b4828a5305343a0e