Add hierarchical structure string printout
Summary:
Adds a simple form of printing the hierarchical structure of the search space; useful in debugging and notebook testing. Looks like this:

{F669905679}

or like this in short form:

{F669905712}

In the future, I would like to use `treelib` for this, but don't want to deal with the Python wheel right now as I don't think pretty-printing this is too important at the moment.

NOTE: Description of plan for this project, with done, in progress, and outstanding items: https://fb.quip.com/vaKQAEaaWJV3

Reviewed By: ldworkin

Differential Revision: D31000945

fbshipit-source-id: 73f3abb567089c3b2c38e9942f8a2dc5e44fae3e