Support search space flattening in `Cast` transform
Summary:
This diff is very much a WiP (docstrings need fixes, included unit-test should be changed to Dev API and moved out of `test_ax_client.py` etc.). Putting it up since seeing this is helpful for review of the previous diffs in the stack.

NOTE: Description of plan for this project, with done, in progress, and outstanding items: https://fb.quip.com/vaKQAEaaWJV3

Reviewed By: ldworkin

Differential Revision: D31001979

fbshipit-source-id: 76808f5642bd6f406a34c926422b3a2f24b090de