Refit ArrayModel on call to get best point
Summary:
In the Service API when the user calls for the best point we use the model from the last good generator run. This is not ideal behavior, because that model is stale in that it doesn't account for the data from the final run.

This diff changes behavior by fitting a new model on the call for best point and surfacing that model's optimization prediction.

NOTE: this behavior is only implemented for ArrayModelBridge. All other models behave as they did previously.

Reviewed By: ldworkin

Differential Revision: D31623503

fbshipit-source-id: 7bec2fc1382689faaf56fc0160c16264419c0151