Fix flakey best_point test
Summary:
Previous iteration of this test had some strange non-deterministic behavior which I now suspect was due to a call to lookup_data in get_best_parameter(). The test would pass in phabricator but not locally (unless focussed) and failed on github.

Over the course of investigating this I also refactored the test to use dev API

Reviewed By: lena-kashtelyan

Differential Revision: D31899019

fbshipit-source-id: 78828b3ff0aaa05f656b8c5bfedc7f8ae9c457d5