adding caption to feature importance plot
Summary:
Not a general purpose solution (i.e., each plot will have to be implemented with this caption separately) but this will allow e.g. AxSweep to caption this advanced plot with a warning/note for users. See the nb in the test plan as well as a followup diff implementing the caption.

https://pxl.cl/1RTPR
Top: without caption
Bottom: with caption (`[ADVANCED]` tag has been added to the title separately)

Reviewed By: lena-kashtelyan

Differential Revision: D31924929

fbshipit-source-id: 45f7c192d6cf9b25aeb15bfd516d8e7641c2620d