Allow LogY transform for absolute constraints
Summary:
It currently isn't possible to use a `LogY` transform on a metric when doing MOO since the objective thresholds are considered outcome constraints. There shouldn't be any issues with applying a `LogY` transform as long as the constraint is absolute and the bound is positive in which case we can just take the log of the bound.

This is also consistent with, e.g., `StandardizeY` which supports absolute constraints, but not relative constraints.

Reviewed By: Balandat

Differential Revision: D31981278

fbshipit-source-id: d9628a2a923a63c33ffc8a3c63ceb18ac9eec03e