Add `ModelSpec` subclass for models specified as factory functions
Summary:
In generation step we support models defined as either members of en enum that extends `ModelsRegistryBase` [1] or a factory function. I'd like to have the following structure for now:

```
dataclass
class GenerationNode:
    model_specs: List[ModelSpec]
    num_trials: int
    ...  # A bunch of other stuff

def fit(...):
    # fits all the model specs

def gen(...):
    # gens from all the model specs
```
```
dataclass
class GenerationStep(GenerationNode):
    model: Union[ModelsRegistryBase, TModelFactoryFunction]
    num_trials: int
    model_kwargs: Dict[str, Any]
    model_gen_kwargs: Dict[str, Any]
    ...  # A bunch of other stuff

    def __post_init__(self):
        # Because we set model specs here, we can directly leverage
        # `super().fit` and `gen` instead of overriding them here
        self.model_specs = [
              ModelSpec(
                  model=self.model,
                  model_kwargs=self.model_kwargs,
                  model_gen_kwargs=self.model_gen_kwargs,
              )
        ]
```

I think the benefits of this structure are:
1) We add a shared and reusable `GenerationNode` abstraction that will contain the bulk of the generation step logic for `fit`, `update`, and `gen` and interface with multiple `ModelSpec`-s to fit/gen from them.
2) We keep a notion of `GenerationStep` to backwards compatibilityand to be able to limit che changes needed to generation strategy at this point; it inherits `fit`, `update`, and `gen` from `GenerationNode`, and its only real difference is that it only has one `ModelSpec` and that model spec is expressed in backwards-compatible way.
    1) All actual logic involved in gen/fit from `GenerationStep` actually delegates to `GenerationNode`.

The only blocker to this structure right now is that `ModelSpec` only accepts model type specified as a member of `Models` enum and not as a factory function (which is allowed by `GenerationStep` and is important functionality, since it's the only way to use the generation strategy with something other than a `Models` member). I think disallowing factory functions in
`GenerationStep` is out of scope for now (although we could add a deprecation warning), so instead my proposed solution is in this diff.

 ---

Reviewed By: ldworkin

Differential Revision: D31996883

fbshipit-source-id: 73e97a0835af3a133d567c2a728ae92931d8e112