fix feature importance caption getting truncated
Summary:
## Relative positions of annotations change from figure to figure

Annotations are currently located on plotly plots using `paper` coordinates, for which 0 corresponds to the minimum of the plot field of view and 1 to the maximum. In order to specify a location beneath the plot, for instance, one must set the y coordinate to a negative number. This is in lieu of a method to specify absolute positions within an entire figure (and not just the plot) or positions relative to axis label positions, etc. One downside is that anything that changes the size of the plot e.g., a long axis label, which squeezes the plot horizontally, or a large number of dimensions, which elongates the plot vertically, will change the relative position of the annotation within the figure. See [link](https://community.plotly.com/t/issues-with-paper-positioned-annotations/4103/4) for some discussion. This can cause part of the annotation to get clipped.

This diff corrects for this by shifting the note a bit left, in proportion to the length of the longest axis label.

Reviewed By: dme65

Differential Revision: D32111258

fbshipit-source-id: afa47ba2dd6927c2e3b9b85e96db51db4111985c