fix in-sample arm visualization with a multi-task model
Summary:
### Problem

Visualization of in-sample arms (e.g. status-quo is None or out-of-design) is incorrect when using multi-task model. This is because we get raw metric values from an arbitrary observation rather the observations from the selected trial (e.g. selecting via FixedFeature) , e.g. as shown in this plot.

The parameter values of status-quo arm is None therefore we use observed raw value rather than model prediction. There are multiple trials and the model is a MTGP.  The `_get_in_sample_arms` uses the obs from latest short-run trial for status-quo while other arms use the MTGP prediction of long-run trial 0. Thus the relative effects are very off.  Please see N1350023 for more plots.
{F680051297}

### Fix
For out-of-design arms, we will check whether the selected trial index is passed via fixed_feature and use the observation from the same trial index.

P.S. another bug is found when data_selector is not None.  We need to modify the t`raining_in_design` list based on the selected obs.

Reviewed By: stevemandala

Differential Revision: D32420148

fbshipit-source-id: 4f4c82d0097f057269703e791e722a5565d05f50