all inputs to `scatter_plot_with_pareto_frontier_plotly` optional except `Y`
Summary: For reporting, we want to handle partial inputs reasonably. Before this input, `scatter_plot_with_pareto_frontier_plotly` required all inputs. Now, only `Y` is required, which enables the metric vs. metric scatterplot. If `Y_pareto` is provided, it is plotted as a step curve, and if `reference_point` is provided, the reference point is plotted with lines connecting it to the pareto curve.

Reviewed By: lena-kashtelyan

Differential Revision: D32494901

fbshipit-source-id: 23bdf9c97b2a91736f80fecde29cf904794f99c4